requirements.txt
numpy
pandas
scipy
scikit-learn
h5py
pyyaml
tqdm
joblib
matplotlib
iterative-stratification
cafaeval